backend/requirements.txt
modal
supermemory
pydantic
httpx
openai
Pillow
